requirements.txt
pyscf
openfermion
torch 
matplotlib
numpy
PyYAML